Added ACL switch to S3 sync command
By adding the acl bucket-owner-full-control switch, the receiving bucket owner has control of the objects placed in their bucket.
When this script is used within the same account this isn't an issue.  In our case, I'm directing the DynamoDB output to a bucket in a different AWS account.  Without this value, a user in the receiving account can't manage the export.  Hope this helps.